refactoring existing rl part in preparation for on_policy methods
